requirements.txt
discord.py[voice]
pynacl
requests
yt-dlp
dotenv